feat: user can delete existing professional experience (fixes #20) and education entries (fixes #24), and the CV preview updates to reflect this. user can also edit existing professional experience (fixes #20) and education entries (fixes #12).
there is now a lot of props being passed downwards, all the way down to the date selector component in both forms. there's also a lot of duplication that #10 refactoring issue should tackle. one option might be to abstract both the forms and previous entries and pass in props that are used conditionally. not sure. might lead to more code instead of less. let's test it out.