feat: setup basic routing with placeholder pages (#48)
* feat(RootLayout): add header as placeholder

* feat(RootLayout): add footer as placeholder

* feat(Header): add links to navigate to other pages

* chore: update formatter rules and apply across codebase

having 2 tab indent instead of 1 tab makes it obvious when my code starts to get a bit too nested - it forces discipline!

* feat: build placeholder pages for all the likely pages (currently), without dynamic subpages, and navigate through them using the navbar in the header

confirmed working! we're using the app-based routing architecture which IIUC is the newer way of doing things in NextJS

* refactor(Header): steps in the app for user to implement details (and thus are pages for routing) are separated out into an array in appSteps.ts

this allows for reusability in the Footer component too, when we navigate between previous and next steps (where available for both)

* feat(appSteps): add helper to get next step's route, and get previous step's route.

return null if there isn't a previous route, a next route (for the relevant helper functions)

* feat(Footer): enable navigation forwards and backwards using 'Previous Step' and 'Next step' Link components from NextJS. use Url type from nextJS for type safety